Rename: modal 폴더 생성 및 모달 관련 컴포넌트 이동
